% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{A2M}
\alias{A2M}
\title{Converts Array to matrix}
\usage{
A2M(
  A,
  n,
  p = if (is.character(n)) {     setdiff(names(dimnames(A), n)) } else {    
    setdiff(1:length(dim(A)), n) }
)
}
\arguments{
\item{A}{An (eventually named) array}

\item{n}{a subvector of 1:length(dim(A)) or names(dimnames(A)).}

\item{p}{a subvector of 1:length(dim(A)) or names(dimnames(A)).}
}
\description{
Converts Array to matrix
}
