% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{extractA}
\alias{extractA}
\title{Extracts dimension from array}
\usage{
extractA(A, a, ...)
}
\arguments{
\item{A}{a named array}

\item{a}{a list of dimensions of A ( a subvector of 1:length(dim(A)) or a subvector of dimnames(A))}

\item{...}{a vector the same length of a of integers. necessarily ...[i]<=dim(A)[a[i]]}
}
\description{
Extracts dimension from array
}
\examples{
A=array(1:(prod(2:4)),2:4);
dimnames(A)<-sapply(dim(A),seq_len)
names(dimnames(A))<-paste0("x",2:4)
extractA(A,integer(0),integer(0));
extractA(A,"x3",2);
extractA(A,c("x4","x3"),1,2);
}
