% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{\%.\%}
\alias{\%.\%}
\title{Define a tensor product}
\usage{
A \%.\% B
}
\arguments{
\item{A}{An array of dimension $\prod_{i\in I_A} a_i$}

\item{B}{An array of dimension $\prod_{i\in I_B} b_i$}

\item{I_A, }{I_B list of subvectors from names(dimnames(A)) and names(dimnames(B))}
}
\value{
C=AB the array of dimension prod_{i\in Ia\cup Ib\cup I}a_1 x ... x a_q x b_1 x ... x b_n  Y[i_1,...,i_q,j_1,...,j_n]=sum_{k_1,...,k_p}(W[i_1,...,i_q,j_1,...,j_n,k_1,...,k_p]*X[j_1,...,i_q,k_1,...,k_p])
}
\description{
Define a tensor product
}
\examples{
A=array(1:(prod(2:6)),2:6);
dimnames(A)<-sapply(dim(A),seq_len)
names(dimnames(A))<-paste0("x",2:6)
B=array(1:(prod(3:7)),3:7);
dimnames(B)<-sapply(dim(B),seq_len)
names(dimnames(B))<-paste0("y",3:7)
I_A=list(c=c("x3","x5"),n=c("x2","x4"),p="x6")
I_B=list(c=c("y3","y5"),q=c("y4","y7"),p="y6")
"\%.\%"(A,B)
"\%.\%"(A,B,I_A,I_B)
W\%.\%.\%.\%t(X);
}

